Silence some warnings + fix fuzzers

+++ src/joedb/journal/Buffered_File.cpp
@@ -126,7 +126,7 @@ namespace joedb
    if (diff)
     return false;
 
-   current += n0;
+   current += int64_t(n0);
   }
 
   return true;
